% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached}
\alias{cached}
\title{Wrap a function to cache its outcome}
\usage{
cached(fn, cache, key)
}
\arguments{
\item{fn}{A function to modify.}

\item{cache}{A simple cache, created through \code{fork::make_cache}}

\item{key}{The key in which to store the cached outcome.}
}
\description{
See \code{fork::make_cache} for details. In particular, note that
this is not a generic memoization cache, it has no invalidation built in,
does not vary the cache by the arguments with which the function is
called... in short, you are responsible for calling
\code{fork::invalidate_cache} whenever necessary.
}
