% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{dotnames}
\alias{dotnames}
\title{Extract the unique column name prefixes from a dataset}
\usage{
dotnames(data, sep = ".", subset = FALSE)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{sep}{The separator between prefix and suffix.}

\item{subset}{Exclude column names that do not have both prefix and suffix.}
}
\value{
A character vector of column name prefixes.
}
\description{
Extract the unique column name prefixes from a dataset
}
