% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{exists}
\alias{exists}
\title{Test for absence of NAs}
\usage{
exists(x)
}
\arguments{
\item{x}{}
}
\value{
A logical value, \code{TRUE} or \code{FALSE}
}
\description{
Essentially \code{!anyNA}
}
\details{
(Can probably be replaced with \verb{compose(!, anyNA)}.)
}
