% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{grid}
\alias{grid}
\title{Create a tibble from all combinations of inputs}
\usage{
grid(..., .name_repair = "check_unique")
}
\arguments{
\item{...}{Name-value pairs. The name will become the column name in the output.}
}
\value{
A tibble
}
\description{
Currently \code{grid} is simply an alias to \code{tidyr::expand_grid}.
It is designed to work with \code{fork::fork} to produce a multiverse specification.
}
