% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{invalidate_cache}
\alias{invalidate_cache}
\title{Invalidate the step cache from a particular step onwards}
\usage{
invalidate_cache(cache, start_at_ix = 1)
}
\arguments{
\item{cache}{A simple step cache as made by \code{make_cache}}

\item{start_at_ix}{The step index from which to invalidate}
}
\value{
The partially invalidated cache.
}
\description{
The step cache is invalidated in-place, not in a copy.
}
