% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{invisibly}
\alias{invisibly}
\title{Wrap a function to run it and pass through its arguments}
\usage{
invisibly(f)
}
\arguments{
\item{f}{The function to modify.}
}
\description{
This is a useful wrapper for functions in compositions (see
\code{purrr:compose}) that is run merely for its side effects.
}
