% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_needed}
\alias{is_needed}
\title{is_needed}
\usage{
is_needed(x)
}
\arguments{
\item{x}{}
}
\value{
A logical value, \code{TRUE} or \code{FALSE}
}
\description{
#' @description
we assume that a step that has multiple parameters, contained within a df-column,
will describe in its first sub-column whether the analysis is needed;
this is naturally what \code{pack_by_sep} will do if a \if{html}{\out{<step>}} variable is defined before
any \if{html}{\out{<step>}}_\if{html}{\out{<parameter>}} variables, but still good to keep in mind;
this function also works with regular columns
}
