% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{link}
\alias{link}
\title{Link steps to scenario columns}
\usage{
link(steps, scenarios, verbose = TRUE)
}
\arguments{
\item{steps}{a named list of functions}

\item{scenarios}{a data frame or tibble with a multiverse specification}

\item{verbose}{TRUE to print links that were established}
}
\description{
Link steps to scenario columns
}
