% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_cache}
\alias{make_cache}
\title{Initialize a simple step cache}
\usage{
make_cache(steps)
}
\arguments{
\item{steps}{a named list, whose names describe the steps to cache}
}
\value{
A cache environment.
}
\description{
We cache the steps of a process, and then for any subsequent
similar process we use the cache of the last step before the process
diverges. Only one process is cached at any time, so this cache relies on a
good ordering of the processes (last step changest fastest) and user-managed
cache invalidation to function effectively.
}
