% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{modifier}
\alias{modifier}
\title{Wrap a function that accepts a list and produces a list, so that it merges input and output}
\usage{
modifier(fn)
}
\arguments{
\item{fn}{A function to modify.}
}
\description{
Wrap a function that accepts a list and produces a list, so that it merges input and output
}
