% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{noop}
\alias{noop}
\title{Return the function arguments as a list.}
\usage{
noop(...)
}
\arguments{
\item{...}{function arguments}
}
\value{
a list of arguments
}
\description{
\code{noop} works similarly to \code{identity} but returns all of its arguments as a list.
}
