% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_by_sep}
\alias{pack_by_sep}
\title{Pack columns that have a prefix.suffix format into df-columns.}
\usage{
pack_by_sep(data, sep = ".")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{sep}{The separator between prefix and suffix.}
}
\value{
A tibble with df-columns.
}
\description{
Pack columns that have a prefix.suffix format into df-columns.
}
