% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{packnames}
\alias{packnames}
\title{Extract the column names of df-columns from a tibble.}
\usage{
packnames(data)
}
\arguments{
\item{data}{A tibble.}
}
\value{
A character vector of column names.
}
\description{
Extract the column names of df-columns from a tibble.
}
