% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{penalize_scenarios}
\alias{penalize_scenarios}
\title{Roughly sort scenarios from preferred to disliked}
\usage{
penalize_scenarios(scenarios, exclude = NULL)
}
\arguments{
\item{scenarios}{A tibble with scenarios, usually generated with \code{grid} and/or \code{fork}}

\item{exclude}{Column names that should not be taken into consideration when rating scenarios.}
}
\value{
A numeric vector of penalties.
}
\description{
A parameter value is taken to be better if it occurs earlier.
Thus, preferences can be changed by changing the order of
parameter values in \code{grid} and \code{fork} calls.
}
