% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{plan_cache}
\alias{plan_cache}
\title{Generate caching hints by analyzing when subsequent scenarios diverge.}
\usage{
plan_cache(scenarios)
}
\arguments{
\item{scenarios}{A tibble with scenarios, usually generated with \code{grid} and/or \code{fork}}
}
\value{
An integer vector with cache invalidation positions.
}
\description{
Generate caching hints by analyzing when subsequent scenarios diverge.
}
