% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{plan_scenarios}
\alias{plan_scenarios}
\title{Annotate scenarios with caching hints and a preference ordering, then pack columns into a more compact format.}
\usage{
plan_scenarios(scenarios, do_not_penalize = NULL)
}
\arguments{
\item{scenarios}{A tibble with scenarios, usually generated with \code{grid} and/or \code{fork}}

\item{do_not_penalize}{}
}
\value{
A tibble with scenarios packed into df-columns and annotated with additional columns.
}
\description{
Annotate scenarios with caching hints and a preference ordering, then pack columns into a more compact format.
}
