% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{replace_na_with_false}
\alias{replace_na_with_false}
\title{Replace NAs with FALSE}
\usage{
replace_na_with_false(data, names = NULL)
}
\arguments{
\item{data}{a data frame}

\item{names}{list of column names}
}
\description{
By default, this function only replaces NAs in "first-level"
columns, e.g. in 'a' but not 'a.b'. It is meant to eventually replace
\code{fork::replace_na_with_falsey}.
}
