% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{replace_na_with_falsey}
\alias{replace_na_with_falsey}
\title{Replace NAs with falsey values, depending on the type of vector}
\usage{
replace_na_with_falsey(data)
}
\arguments{
\item{data}{A vector.}
}
\description{
To state the obvious, it is not in general a good idea to replace NA values
with FALSE or its equivalents, but in the context of a multiverse scenario,
NA is meant to convey that the parameter does not apply and so its value is
irrelevant.

Replacing NA with falsey values in scenarios allows for much easier filtering:
\itemize{
\item \verb{is.na(x) | !x)} can be replaced with \code{!x}
\item \code{!is.na(x) & x == 'abc'} can be replaced with \code{x == 'abc'}
}

\code{replace_na_with_falsey} must be called explicitly, it is not a part of the
behavior of \code{fork}
}
