% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{reroute}
\alias{reroute}
\title{Reroute named arguments (or a component thereof) to other named arguments}
\usage{
reroute(.fn, ...)
}
\arguments{
\item{...}{a mapping of destination=c('path', 'to', 'source') as used by \code{purrr:pluck}}

\item{fn}{a function whose arguments to intercept}
}
\description{
Reroute named arguments (or a component thereof) to other named arguments
}
\examples{
invalidate_cache_by_ix <- reroute(invalidate_cache, start_at_ix=c('input', 'start_at_ix'))
}
