% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{rethrow_with_args}
\alias{rethrow_with_args}
\title{Modify a function to rethrow error conditions with function name and arguments}
\usage{
rethrow_with_args(fn, name)
}
\arguments{
\item{fn}{A function to be modified.}

\item{name}{Name of the function.}
}
\description{
Modify a function to rethrow error conditions with function name and arguments
}
