% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{rethrow_with_arguments}
\alias{rethrow_with_arguments}
\title{Modify a function to rethrow error conditions with function name and arguments}
\usage{
rethrow_with_arguments(fn, name = NULL)
}
\arguments{
\item{fn}{A function to be modified.}

\item{name}{Name of the function.}
}
\description{
Modify a function to rethrow error conditions with function name and arguments
}
