% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{row_modify}
\alias{row_modify}
\title{Modify the rows of a data frame, one at a time, and bind the results with the
original rows}
\usage{
row_modify(data, fn)
}
\arguments{
\item{data}{a data frame}

\item{fn}{a function that takes a one-row data frame and returns a one-row data frame}
}
\description{
\code{row_modify(...)} is very similar to \code{dplyr::rowwise() |> dplyr::mutate(...)} and if possible you should prefer to use that workflow,
however \code{fork::row_modify} works on the actual row instead of an indirect
representation of the row (\code{.data}), which leads to fewer surprises when all
columns of the row are relevant to the modification.
}
