% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{skippable}
\alias{skippable}
\title{Only execute a function if a particular argument is truthy}
\usage{
skippable(fn, path, otherwise)
}
\arguments{
\item{fn}{A function to modify}

\item{path}{The name of an argument (or component thereof)}

\item{otherwise}{The return value if the function is not executed}
}
\description{
Only execute a function if a particular argument is truthy
}
