% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{splat}
\alias{splat}
\title{Return the function arguments as a list.}
\usage{
splat(...)
}
\arguments{
\item{...}{function arguments}
}
\value{
a list of arguments
}
\description{
\code{splat} works similarly to \code{identity} but returns all of its arguments as a list, not just the first one.
}
