% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{step_through}
\alias{step_through}
\title{Create a \code{step} function that is aware of the cache and the active scenario.}
\usage{
step_through(scenario, cache)
}
\arguments{
\item{scenario}{One scenario, usually from a tibble of scenarios generated with \code{grid} and \code{fork}.}

\item{cache}{A step cache as generated by \code{make_cache}.}
}
\value{
A function with signature \code{step(name, fn)}
}
\description{
The \code{step(name, fn)} function created by this factory will query
the cache to see whether execution is necessary, and if it is, pass the
output from the previous step to \code{fn}, which will in turn be cached.
}
\examples{
scenarios <- expand_grid(add=1:3, multiply=1:3)
cache <- make_cache(scenarios)
output <- lapply(1:nrow(scenarios), function(i) {
  for (i in 1:nrow(scenarios)) {
    step <- step_through(scenario, cache)
    step('add', function(input) {
      10 + scenario$add
    })
    step('multiply', function(input) {
      input * scenario$multiply
    })
    return(cache$multiply)
  }
}
}
