% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack_by_sep}
\alias{unpack_by_sep}
\title{Unpack df-columns into regular columns with a prefix.suffix naming scheme}
\usage{
unpack_by_sep(data, sep = ".")
}
\arguments{
\item{data}{A tibble.}

\item{sep}{The separator between prefix and suffix.}
}
\value{
A tibble without df-columns.
}
\description{
Unpack df-columns into regular columns with a prefix.suffix naming scheme
}
