% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{unsplat}
\alias{unsplat}
\title{Wrap a function that takes multiple arguments to instead take a list of arguments}
\usage{
unsplat(fn)
}
\arguments{
\item{fn}{A function to modify.}
}
\description{
Wrap a function that takes multiple arguments to instead take a list of arguments
}
