% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{verify_steps}
\alias{verify_steps}
\title{Verify whether each fork in a multiverse specification has a corresponding step function.}
\usage{
verify_steps(steps, scenarios)
}
\arguments{
\item{steps}{a named list of functions}

\item{scenarios}{a data frame with a multiverse specification, such as produced by \code{fork::fork}}
}
\description{
Verify whether each fork in a multiverse specification has a corresponding step function.
}
