% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{fork}
\alias{fork}
\title{Create a tibble from partial combinations of inputs}
\usage{
fork(.left, .sep = ".", ...)
}
\arguments{
\item{.left}{A tibble as created with \code{grid}, \code{tibble}, \code{tribble} or a previous call to \code{fork}}

\item{...}{Name-value pairs to produce a righthand tibble that will be crossed with the lefthand one.}
}
\value{
A tibble.
}
\description{
Create a tibble from partial combinations of inputs
}
\examples{
tibble(analysis = c("lm", "glm")) \%>\% fork(analysis = "glm", family = c("logit", "probit"))
}
