% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_cache}
\alias{make_cache}
\title{Initialize a simple step cache}
\usage{
make_cache(scenarios)
}
\arguments{
\item{scenarios}{}
}
\value{
A cache environment.
}
\description{
Initialize a simple step cache
}
